+++ ext/dmddevice/pinDMDv3/dmddevice_dll/dmddevice.cpp
@@ -81,10 +81,10 @@ DMDDEV void Set_16_Colors_Palette(rgb24 *color)
 
 DMDDEV void Render_4_Shades(UINT16 width, UINT16 height, UINT8 *currbuffer)
 {
-	if (!memcmp(oldbuffer, currbuffer, width*height)) //check if same frame again
+	if (!memcmp(oldbuffer, currbuffer, (size_t)width*height)) //check if same frame again
 		return;
 
-	memcpy(oldbuffer, currbuffer, width*height);
+	memcpy(oldbuffer, currbuffer, (size_t)width*height);
 
 	if (isOpen) {
 		renderDMDFrame(0x00000000080, (UINT8) width, (UINT8) height, currbuffer, 0x00); 
@@ -93,10 +93,10 @@ DMDDEV void Render_4_Shades(UINT16 width, UINT16 height, UINT8 *currbuffer)
 
 DMDDEV void Render_16_Shades(UINT16 width, UINT16 height, UINT8 *currbuffer) 
 {
-	if (!memcmp(oldbuffer, currbuffer, width*height)) //check if same frame again
+	if (!memcmp(oldbuffer, currbuffer, (size_t)width*height)) //check if same frame again
 		return;
 
-	memcpy(oldbuffer, currbuffer, width*height);
+	memcpy(oldbuffer, currbuffer, (size_t)width*height);
 
 	if (isOpen) {
 		render16ShadeFrame(currbuffer);
@@ -190,10 +190,10 @@ DMDDEV void Render_PM_Alphanumeric_Frame(layout_t layout, const UINT16 *const se
 
 DMDDEV void Render_RGB24(UINT16 width, UINT16 height, rgb24 *currbuffer)
 {
-	if (!memcmp(oldbuffer, currbuffer, width*height)) //check if same frame again
+	if (!memcmp(oldbuffer, currbuffer, (size_t)width*height)) //check if same frame again
 		return;
 
-	memcpy(oldbuffer, currbuffer, width*height);
+	memcpy(oldbuffer, currbuffer, (size_t)width*height);
 
 	if (!isOpen && !useRGB) {
 		tPMoptions Options;
